Implement ABAC to reistrict reports
