+++ studio/components/to-be-cleaned/ProductEmptyState.tsx
@@ -1,5 +1,6 @@
 import { Button, Typography } from '@supabase/ui'
 import { FC, ReactNode } from 'react'
+import * as Tooltip from '@radix-ui/react-tooltip'
 
 interface Props {
   title?: string
@@ -9,6 +10,7 @@ interface Props {
   infoButtonUrl?: string
   onClickCta?: () => void
   disabled?: boolean
+  disabledMessage?: string
 }
 
 const ProductEmptyState: FC<Props> = ({
@@ -19,6 +21,7 @@ const ProductEmptyState: FC<Props> = ({
   infoButtonUrl = '',
   onClickCta = () => {},
   disabled = false,
+  disabledMessage = '',
 }) => (
   <div className="flex items-center justify-center w-full h-full">
     <div className="flex space-x-4 p-6 rounded border border-panel-border-light dark:border-panel-border-dark bg-panel-body-light dark:bg-panel-body-dark shadow-md">
@@ -28,9 +31,26 @@ const ProductEmptyState: FC<Props> = ({
           <Typography.Title level={5}>{title}</Typography.Title>
           <div className="flex flex-col space-y-2">{children}</div>
           <div className="flex items-center space-x-2">
-            <Button type="primary" onClick={onClickCta} disabled={disabled}>
-              {ctaButtonLabel}
-            </Button>
+            <Tooltip.Root delayDuration={0}>
+              <Tooltip.Trigger>
+                <Button type="primary" onClick={onClickCta} disabled={disabled}>
+                  {ctaButtonLabel}
+                </Button>
+              </Tooltip.Trigger>
+              {disabled && disabledMessage && (
+                <Tooltip.Content side="bottom">
+                  <Tooltip.Arrow className="radix-tooltip-arrow" />
+                  <div
+                    className={[
+                      'bg-scale-100 rounded py-1 px-2 leading-none shadow',
+                      'border-scale-200 border',
+                    ].join(' ')}
+                  >
+                    <span className="text-scale-1200 text-xs">{disabledMessage}</span>
+                  </div>
+                </Tooltip.Content>
+              )}
+            </Tooltip.Root>
             {infoButtonUrl && infoButtonLabel ? (
               <Button type="default">
                 <a target="_blank" href={infoButtonUrl}>
