+++ studio/components/to-be-cleaned/Reports/Reports.utils.ts
@@ -1,6 +1,5 @@
 import { useProjectContentStore } from 'stores/projectContentStore'
 import { NEW_REPORT_SKELETON } from './Reports.constants'
-import toast from 'react-hot-toast'
 
 /*
  * createReport()
